feat: updated loadLogs to match other loading functions
