+++ src/stores/logStore.js
@@ -11,3 +11,26 @@ export const filterDropdownState = writable({}); // State for filter dropdowns
 export const sortOrder = writable('asc'); // Current sort order for logs (asc or desc)
 
 export const logColumns = writable([]); // Array of column names to display in the table
+
+export async function loadLogs() {
+  try {
+    const res = await fetch('/api/log');
+    if (!res.ok) throw new Error('Failed to fetch log file');
+
+    const text = await res.text();
+
+    if (!text.trim()) {
+      console.log('Loaded logs: empty response');
+      return [];
+    }
+
+    const parsedLogs = text
+      .split('\n')
+      .filter(line => line.trim())
+      .map(line => JSON.parse(line));
+  
+    logs.set(parsedLogs);
+  } catch (err) {
+    console.error('Fetch error:', err);
+  }
+}
